+++ src/server/components/auth/router.ts
@@ -1,0 +1,14 @@
+import express from "express";
+import {
+  authController,
+  getLoginController,
+  postLoginController,
+  getLogoutController,
+} from "./controllers";
+import { authRoutes } from "./constants";
+
+export const authRouter = express.Router();
+
+authRouter.get(authRoutes.login, getLoginController, authController);
+authRouter.post(authRoutes.login, postLoginController);
+authRouter.get(authRoutes.logout, getLogoutController);
