+++ src/server/components/dashboard/index.ts
@@ -1,12 +1,1 @@
-export {
-  startRouteController,
-  usersListController,
-  usersEditController,
-  listsController,
-  listsEditController,
-  listsItemsController,
-  listItemsApproveController,
-  listItemsPublishController,
-} from "./dashboard";
-export { dashboardRoutes } from "./routes";
-export { dashboardRouter } from "./router";
+export { dashboardRouter, dashboardRoutes } from "./router";
