+++ src/server/components/feedback/index.ts
@@ -1,2 +1,1 @@
-export { feedbackRoutes } from "./routes";
-export { feedbackIngest } from "./feedback";
+export { feedbackRouter } from "./router";
