--- src/server/components/feedback/routes.ts
@@ -1,3 +1,0 @@
-export const feedbackRoutes = {
-  post: "/feedback"
-}
