+++ src/server/components/lists/constants.ts
@@ -1,6 +1,6 @@
 import { countriesList } from "server/services/metadata";
 import { ServiceType } from "server/models/types";
-import { listsRoutes } from "./routes";
+import { listsRoutes } from "./router";
 
 export const DEFAULT_VIEW_PROPS = {
   countriesList,
