+++ src/server/components/lists/index.ts
@@ -1,12 +1,3 @@
-export {
-  listsGetController,
-  listsPostController,
-  listsResultsController,
-  listsGetPrivateBetaPage,
-  listsDataIngestionController,
-  listsConfirmApplicationController,
-} from "./lists";
-
-export { listsRoutes } from "./routes";
+export { listsRoutes, listsRouter } from "./router";
 export { getServiceLabel, createListSearchBaseLink } from "./helpers"
 export * from "./types";
