+++ src/server/components/lists/router.ts
@@ -1,16 +1,24 @@
 /* eslint-disable @typescript-eslint/no-misused-promises */
 import express from "express";
 import {
-  listsRoutes,
-  listsFeedbackPost,
   listsGetController,
-  listsFeedbackSuccess,
   listsPostController,
   listsResultsController,
   listsGetPrivateBetaPage,
   listsDataIngestionController,
   listsConfirmApplicationController,
-} from "server/components/lists";
+} from "./lists";
+
+export const listsRoutes = {
+  start: "/",
+  finder: "/find",
+  results: "/results",
+  feedback: "/feedback",
+  feedbackSuccess: "/feedback/success",
+  privateBeta: "/private-beta",
+  formRunnerWebhook: "/ingest/:serviceType",
+  confirmApplication: "/confirm/:reference",
+};
 
 export const listsRouter = express.Router();
 
@@ -20,5 +28,3 @@ listsRouter.get(listsRoutes.results, listsResultsController);
 listsRouter.post(listsRoutes.formRunnerWebhook, listsDataIngestionController);
 listsRouter.get(listsRoutes.confirmApplication, listsConfirmApplicationController);
 listsRouter.get(listsRoutes.privateBeta, listsGetPrivateBetaPage);
-listsRouter.post(listsRoutes.feedback, listsFeedbackPost);
-listsRouter.get(listsRoutes.feedbackSuccess, listsFeedbackSuccess);
