--- src/server/components/lists/routes.ts
@@ -1,10 +1,0 @@
-export const listsRoutes = {
-  start: "/",
-  finder: "/find",
-  results: "/results",
-  feedback: "/feedback",
-  feedbackSuccess: "/feedback/success",
-  privateBeta: "/private-beta",
-  formRunnerWebhook: "/ingest/:serviceType",
-  confirmApplication: "/confirm/:reference",
-};
