+++ src/server/components/sitemap/index.ts
@@ -1,2 +1,1 @@
-export { sitemapController } from "./sitemap";
 export { sitemapRouter } from "./router";
